Avoid interference from text body parser
This this router is often added at the root of the express server, the `text()` body parser would be executed for every other route. In my codebase, I have routes that require the body to be unprocessed. This change should add the `text()` body parser to the relevant routes, without interfering with the rest of the express app.